chore: clarify package.json
