+++ .gitignore
@@ -32,10 +32,4 @@ yarn-error.log*
 
 # server
 server/**/$*.ts
-server/**/*.db
 server/index.js
-server/snapshot.js
-server/.upload
-server/public/icons/*
-!server/public/icons/dammy.svg
-!server/**/.env
