+++ server/templates/client/next/package.json
@@ -5,38 +5,34 @@
   "scripts": {
     "dev": "<% if (orm === 'prisma') { %>npm run migrate:up && <% } %>run-p dev:*",
     "dev:client": "next dev -p <%= clientPort %>",
-    "dev:server": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/webpack --watch --mode=development",
+    "dev:server": "npm run dev --prefix server",
     "dev:aspida": "aspida --watch",
-    "dev:frourio": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/frourio<% if (server === 'express') { %>-express<% } %> --watch<% if (orm === 'prisma') { %>",
-    "dev:prisma": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma generate --watch<% } else if (orm === 'typeorm') { %>",
-    "dev:typeormer": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/typeormer --watch<% } %>",
     "build": "run-p build:client build:server",
     "build:client": "aspida && next build<% if (building === 'static') { %> && next export<% } %>",
-    "build:server": "<% if (orm === 'prisma') { %>npm run migrate:up && <% } %>npm run build:frourio && cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/webpack --mode=production",
-    "build:frourio": "cd server && <% if (orm === 'prisma') { %>cross-env DUMMY_ENV=forWin node_modules/.bin/prisma generate && <% } else if (orm === 'typeorm') { %>cross-env DUMMY_ENV=forWin node_modules/.bin/typeormer && <% } %>cross-env DUMMY_ENV=forWin node_modules/.bin/frourio<% if (server === 'express') { %>-express<% } %>",
-    "build:types": "aspida && npm run build:frourio",
+    "build:server": "npm run build --prefix server",
+    "build:types": "aspida && npm run build:frourio --prefix server",
     "lint": "eslint --ext .ts,.js,.tsx .",
     "lint:fix": "npm run lint -- --fix<% if (orm === 'prisma') { %>",
     "migrate": "npm run migrate:save && npm run migrate:up",
-    "migrate:save": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma migrate save --create-db --experimental",
-    "migrate:up": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma migrate up --create-db --experimental",
-    "migrate:down": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma migrate down --experimental<% } else if (orm === 'typeorm') { %>",
-    "migration:generate": "cd server && ts-node cross-env DUMMY_ENV=forWin node_modules/.bin/typeorm migration:generate -n Task",
-    "migration:run": "cd server && ts-node cross-env DUMMY_ENV=forWin node_modules/.bin/typeorm migration:run",
-    "migration:revert": "cd server && ts-node cross-env DUMMY_ENV=forWin node_modules/.bin/typeorm migration:revert<% } %>",
+    "migrate:save": "npm run migrate:save --prefix server",
+    "migrate:up": "npm run migrate:up --prefix server",
+    "migrate:down": "npm run migrate:down --prefix server<% } else if (orm === 'typeorm') { %>",
+    "migration:generate": "npm run migration:generate --prefix server",
+    "migration:run": "npm run migration:run --prefix server",
+    "migration:revert": "npm run migration:revert --prefix server<% } %>",
     "start": "run-p start:*",
     "start:client": "next start",
-    "start:server": "cd server && cross-env NODE_ENV=production node index.js<% if (testing === 'jest') { %>",
+    "start:server": "npm start --prefix server<% if (testing === 'jest') { %>",
     "test": "npm run build:types && jest<% } %>",
     "typecheck": "npm run build:types && tsc --noEmit && tsc --noEmit -p server"
   },
   "dependencies": {
     "<% if (aspida === 'axios') { %>@aspida/axios": "^0.12.2",
     "<% } else { %>@aspida/fetch": "^0.10.2",
-    "<% } %>@aspida/swr": "^0.2.2",
-    "<% if (aspida === 'axios') { %>axios": "^0.21.0",
-    "<% } %>class-validator": "^0.12.2",
-    "next": "^10.0.0",
+    "<% } %>@aspida/swr": "^0.2.2<% if (aspida === 'axios') { %>",
+    "axios": "^0.21.0<% } %>",
+    "class-validator": "^0.12.2",
+    "next": "^10.0.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "swr": "^0.3.8"
@@ -44,20 +40,20 @@
   "devDependencies": {
     "<% if (testing === 'jest') { %>@types/jest": "^26.0.14",
     "<% } %>@types/node": "^14.11.2",
-    "@types/react": "^16.9.49",
-    "@typescript-eslint/eslint-plugin": "^4.4.1",
-    "@typescript-eslint/parser": "^4.4.1",
+    "@types/react": "^16.9.56",
+    "@typescript-eslint/eslint-plugin": "^4.6.1",
+    "@typescript-eslint/parser": "^4.6.1",
     "cross-env": "^7.0.2",
     "dotenv": "^8.2.0",
-    "eslint": "^7.11.0",
-    "eslint-config-prettier": "^6.13.0",
+    "eslint": "^7.13.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.4",
-    "<% if (testing === 'jest') { %>jest": "^26.4.2",
-    "<% } %>npm-run-all": "^4.1.5",
-    "prettier": "^2.1.2",
-    "<% if (testing === 'jest') { %>ts-jest": "^26.4.1",
-    "<% } %>ts-loader": "^8.0.5",
+    "eslint-plugin-react": "^7.21.5<% if (testing === 'jest') { %>",
+    "jest": "^26.6.3<% } %>",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.1.2<% if (testing === 'jest') { %>",
+    "ts-jest": "^26.4.3<% } %>",
+    "ts-loader": "^8.0.10",
     "typescript": "^4.0.5"
   }
 }
