+++ server/templates/client/nuxt/package.json
@@ -5,30 +5,26 @@
   "scripts": {
     "dev": "<% if (orm === 'prisma') { %>npm run migrate:up && <% } %>run-p dev:*",
     "dev:client": "cross-env NUXT_PORT=<%= clientPort %> nuxt-ts",
-    "dev:server": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/webpack --watch --mode=development",
+    "dev:server": "npm run dev --prefix server",
     "dev:aspida": "aspida --watch",
-    "dev:frourio": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/frourio<% if (server === 'express') { %>-express<% } %> --watch<% if (orm === 'prisma') { %>",
-    "dev:prisma": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma generate --watch<% } else if (orm === 'typeorm') { %>",
-    "dev:typeormer": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/typeormer --watch<% } %>",
     "build": "run-p build:client build:server",
     "build:client": "aspida && nuxt-ts generate",
-    "build:server": "<% if (orm === 'prisma') { %>npm run migrate:up && <% } %>npm run build:frourio && cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/webpack --mode=production",
-    "build:frourio": "cd server && <% if (orm === 'prisma') { %>cross-env DUMMY_ENV=forWin node_modules/.bin/prisma generate && <% } else if (orm === 'typeorm') { %>cross-env DUMMY_ENV=forWin node_modules/.bin/typeormer && <% } %>cross-env DUMMY_ENV=forWin node_modules/.bin/frourio<% if (server === 'express') { %>-express<% } %>",
-    "build:types": "aspida && npm run build:frourio",
+    "build:server": "npm run build --prefix server",
+    "build:types": "aspida && npm run build:frourio --prefix server",
     "lint": "npm run lint:js && npm run lint:style",
     "lint:js": "eslint --ext .ts,.js,.vue .",
     "lint:style": "stylelint **/*.{vue,css} --ignore-path .gitignore",
     "lint:fix": "npm run lint:js -- --fix && npm run lint:style -- --fix<% if (orm === 'prisma') { %>",
     "migrate": "npm run migrate:save && npm run migrate:up",
-    "migrate:save": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma migrate save --create-db --experimental",
-    "migrate:up": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma migrate up --create-db --experimental",
-    "migrate:down": "cd server && cross-env DUMMY_ENV=forWin node_modules/.bin/prisma migrate down --experimental<% } else if (orm === 'typeorm') { %>",
-    "migration:generate": "cd server && ts-node cross-env DUMMY_ENV=forWin node_modules/.bin/typeorm migration:generate -n Task",
-    "migration:run": "cd server && ts-node cross-env DUMMY_ENV=forWin node_modules/.bin/typeorm migration:run",
-    "migration:revert": "cd server && ts-node cross-env DUMMY_ENV=forWin node_modules/.bin/typeorm migration:revert<% } %>",
+    "migrate:save": "npm run migrate:save --prefix server",
+    "migrate:up": "npm run migrate:up --prefix server",
+    "migrate:down": "npm run migrate:down --prefix server<% } else if (orm === 'typeorm') { %>",
+    "migration:generate": "npm run migration:generate --prefix server",
+    "migration:run": "npm run migration:run --prefix server",
+    "migration:revert": "npm run migration:revert --prefix server<% } %>",
     "start": "run-p start:*",
     "start:client": "nuxt-ts <%= target === 'static' ? 'serve' : 'start' %>",
-    "start:server": "cd server && cross-env NODE_ENV=production node index.js<% if (testing === 'jest') { %>",
+    "start:server": "npm start --prefix server<% if (testing === 'jest') { %>",
     "test": "npm run build:types && jest<% } %>",
     "typecheck": "npm run build:types && tsc --noEmit && tsc --noEmit -p server"
   },
@@ -45,35 +41,35 @@
     "<% if (aspida === 'axios') { %>@aspida/axios": "^0.12.2",
     "<% } else { %>@aspida/fetch": "^0.10.2",
     "@aspida/node-fetch": "^0.9.2",
-    "<% } %>@nuxt/typescript-runtime": "^2.0.0",
-    "<% if (aspida === 'axios') { %>@nuxtjs/axios": "^5.12.2",
-    "<% } %>class-validator": "^0.12.2",
+    "<% } %>@nuxt/typescript-runtime": "^2.0.0<% if (aspida === 'axios') { %>",
+    "@nuxtjs/axios": "^5.12.2<% } %>",
+    "class-validator": "^0.12.2",
     "nuxt": "2.14.7"
   },
   "devDependencies": {
-    "@nuxt/types": "^2.14.6",
+    "@nuxt/types": "^2.14.7",
     "@nuxt/typescript-build": "^2.0.3",
-    "@nuxtjs/eslint-config": "^3.1.0",
-    "@nuxtjs/eslint-config-typescript": "^3.0.0",
-    "@nuxtjs/eslint-module": "^2.0.0",
+    "@nuxtjs/eslint-config": "^4.0.0",
+    "@nuxtjs/eslint-config-typescript": "^4.0.0",
+    "@nuxtjs/eslint-module": "^3.0.0",
     "@nuxtjs/stylelint-module": "^4.0.0<% if (testing === 'jest') { %>",
-    "@types/jest": "^26.0.14<% } %><% if (aspida === 'fetch') { %>",
+    "@types/jest": "^26.0.15<% } %><% if (aspida === 'fetch') { %>",
     "@types/node-fetch": "^2.5.7<% } %>",
     "babel-eslint": "^10.1.0",
     "cross-env": "^7.0.2",
     "dotenv": "^8.2.0",
-    "eslint": "^7.10.0",
-    "eslint-config-prettier": "^6.12.0",
+    "eslint": "^7.13.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-nuxt": "^1.0.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "husky": "^4.3.0",
-    "<% if (testing === 'jest') { %>jest": "^26.4.2",
-    "<% } %>lint-staged": "^10.4.0",
+    "husky": "^4.3.0<% if (testing === 'jest') { %>",
+    "jest": "^26.6.3<% } %>",
+    "lint-staged": "^10.5.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.1.2",
-    "stylelint": "^13.7.2",
-    "<% if (testing === 'jest') { %>ts-jest": "^26.4.1",
-    "<% } %>ts-loader": "^8.0.9",
+    "stylelint": "^13.7.2<% if (testing === 'jest') { %>",
+    "ts-jest": "^26.4.3<% } %>",
+    "ts-loader": "^8.0.10",
     "typescript": "^4.0.5"
   }
 }
