+++ server/templates/server/package.json
@@ -8,7 +8,8 @@
     "dev:frourio": "frourio<% if (server === 'express') { %>-express<% } %> --watch<% if (orm === 'prisma') { %>",
     "dev:prisma": "prisma generate --watch<% } else if (orm === 'typeorm') { %>",
     "dev:typeormer": "typeormer --watch<% } %>",
-    "build": "<% if (orm === 'prisma') { %>npm run migrate:up && prisma generate && <% } else if (orm === 'typeorm') { %>typeormer && <% } %>frourio<% if (server === 'express') { %>-express<% } %> && webpack --mode=production<% if (orm === 'prisma') { %>",
+    "build": "npm run build:frourio && webpack --mode=production",
+    "build:frourio": "<% if (orm === 'prisma') { %>npm run migrate:up && prisma generate && <% } else if (orm === 'typeorm') { %>typeormer && <% } %>frourio<% if (server === 'express') { %>-express<% } %><% if (orm === 'prisma') { %>",
     "migrate": "npm run migrate:save && npm run migrate:up",
     "migrate:save": "prisma migrate save --create-db --experimental",
     "migrate:up": "prisma migrate up --create-db --experimental",
@@ -53,7 +54,7 @@
     "frourio<% if (server === 'express') { %>-express<% } %>": "^0.19.1",
     "nodemon-webpack-plugin": "^4.3.2",
     "npm-run-all": "^4.1.5",
-    "ts-loader": "^8.0.9",
+    "ts-loader": "^8.0.10",
     "ts-node": "^9.0.0",
     "tsconfig-paths-webpack-plugin": "^3.3.0",
     "typescript": "^4.0.5",
